Clean up map updates
